ministry-billing-report/requirements.txt
requests>=2.18.4,<3.0
flake8==3.9.2
black==21.6b0
python-dotenv==0.17.1
s3==3.0.0
pyyaml==5.4.1